What I Have So Far
This is what I have of the project so far. I tried to create it in
Visual Studio but since I did not have extensive knowledge of Visual
Studio and Visual Studio was not designed with games in mind, I scrapped
it. I'm now trying to learn how to use Unity and try to replicate what I
had in the Visual Studio version. It's not going to be easy but I'll do
my best.